requirements-dev.txt
#######################################################
### Include                                         ###
#######################################################

-r requirements.txt

#######################################################
### Development                                     ###
#######################################################

django-debug-toolbar>=0.9.4
django-template-debug>=0.3.3

ipython>=2.0.0
ipdb>=0.8
readline>=6.2.4.1
gnureadline>=6.3.3
factory_boy==2.4.1
fake-factory==0.4.2
mock==1.0.1

#######################################################
### Deployment                                      ###
#######################################################

#collectfast

#closure==20121212  # Second to latest version, because mac doesnt have java 52 yet (need to figure out how to update)
#yuicompressor>=2.4.8
